GA5

+++ portal-impl/src/com/liferay/portal/lar/backgroundtask/LayoutRemoteStagingBackgroundTaskExecutor.java
@@ -18,9 +18,13 @@ import com.liferay.portal.NoSuchLayoutException;
 import com.liferay.portal.RemoteExportException;
 import com.liferay.portal.kernel.backgroundtask.BackgroundTaskResult;
 import com.liferay.portal.kernel.lar.ExportImportHelperUtil;
+import com.liferay.portal.kernel.lar.ExportImportThreadLocal;
 import com.liferay.portal.kernel.lar.MissingReferences;
 import com.liferay.portal.kernel.lar.PortletDataHandlerKeys;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.staging.StagingUtil;
+import com.liferay.portal.kernel.util.DateRange;
 import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.MapUtil;
@@ -82,6 +86,8 @@ public class LayoutRemoteStagingBackgroundTaskExecutor
 		MissingReferences missingReferences = null;
 
 		try {
+			ExportImportThreadLocal.setLayoutStagingInProcess(true);
+
 			file = exportLayoutsAsFile(
 				sourceGroupId, privateLayout, layoutIdMap, parameterMap,
 				remoteGroupId, startDate, endDate, httpPrincipal);
@@ -139,19 +145,38 @@ public class LayoutRemoteStagingBackgroundTaskExecutor
 				parameterMap, PortletDataHandlerKeys.UPDATE_LAST_PUBLISH_DATE);
 
 			if (updateLastPublishDate) {
+				DateRange dateRange = new DateRange(startDate, endDate);
+
 				StagingUtil.updateLastPublishDate(
-					sourceGroupId, privateLayout, endDate);
+					sourceGroupId, privateLayout, dateRange, endDate);
 			}
 		}
-		finally {
-			StreamUtil.cleanUp(fileInputStream);
+		catch (Exception e) {
+			if (PropsValues.STAGING_DELETE_TEMP_LAR_ON_FAILURE) {
+				FileUtil.delete(file);
+			}
+			else if ((file != null) && _log.isErrorEnabled()) {
+				_log.error("Kept temporary LAR file " + file.getAbsolutePath());
+			}
 
-			FileUtil.delete(file);
+			throw e;
+		}
+		finally {
+			ExportImportThreadLocal.setLayoutStagingInProcess(false);
+
+			StreamUtil.cleanUp(fileInputStream);
 
 			if (stagingRequestId > 0) {
 				StagingServiceHttp.cleanUpStagingRequest(
 					httpPrincipal, stagingRequestId);
 			}
+		}
+
+		if (PropsValues.STAGING_DELETE_TEMP_LAR_ON_SUCCESS) {
+			FileUtil.delete(file);
+		}
+		else if ((file != null) && _log.isDebugEnabled()) {
+			_log.debug("Kept temporary LAR file " + file.getAbsolutePath());
 		}
 
 		return processMissingReferences(backgroundTask, missingReferences);
@@ -248,4 +273,6 @@ public class LayoutRemoteStagingBackgroundTaskExecutor
 		return missingRemoteParentLayouts;
 	}
 
+	private static Log _log = LogFactoryUtil.getLog(
+		LayoutRemoteStagingBackgroundTaskExecutor.class);
 }
