+++ portal-impl/src/com/liferay/portlet/rolesadmin/lar/RolesAdminPortletDataHandler.java
@@ -78,7 +78,7 @@ public class RolesAdminPortletDataHandler extends BasePortletDataHandler {
 			portletDataContext.getCompanyId());
 
 		for (Role role : roles) {
-			if (!PortalUtil.isSystemRole(role.getName())) {
+			if (!PortalUtil.isSystemRole(role.getName()) && !role.isTeam()) {
 				RoleLocalServiceUtil.deleteRole(role);
 			}
 		}
